fixed #1712: Adding default dpi for marmalade and mac since i don't know how to get DPI for these platforms.

+++ cocos2dx/platform/blackberry/CCDevice.cpp
@@ -6,7 +6,7 @@
 
 NS_CC_BEGIN
 
-float CCDevice::getDPI()
+int CCDevice::getDPI()
 {
     static int dpi = -1;
     if (dpi == -1)
